Set `features_default_export_path` to `sites/all/modules/custom`.
The default value in features is `sites/all/modules`.

So each time a developer creates a new feature one must manually add
`sites/all/modules/custom` if generating the feature from the web UI.

That leaves an entry in the `.info` file:
```
project path = sites/all/modules/custom
```
The entry is not really needed and makes it harder to actually move a
module to a different location.

So let's just configure a default export path for features and never be
bother with it again.